Add files via upload
In this component , You will find on selection of role Data will come from respective object as what is the key roles and responsibility and respective weightage  is for employee and manager will give his or her ratings and comments. based on the type of user logged in the self rating and manager rating blocks are editable.